+++ src/metrics/index.ts
@@ -1,0 +1,2 @@
+export { default as MetricsManager } from './Manager'
+export { default as MetricsWorker } from './Worker'